refactor integrations folder
