+++ nanome/_internal/integration/__init__.py
@@ -1,4 +1,4 @@
 from . import *
 # classes
 # folders
-from . import serialization
+from . import serializers
