--- nanome/_internal/integration/serialization/export_file.py
@@ -1,20 +1,0 @@
-from nanome._internal.util.type_serializers import TypeSerializer, StringSerializer
-
-
-class _ExportFile(TypeSerializer):
-    _String = StringSerializer()
-
-    def version(self):
-        return 0
-
-    def name(self):
-        return "ExportFile"
-
-    def serialize(self, version, value, context):
-        context.write_bool(value)
-
-    def deserialize(self, version, context):
-        location = context.read_using_serializer(_ExportFile._String)
-        filename = context.read_using_serializer(_ExportFile._String)
-        data = context.read_byte_array()
-        return (location, filename, data)
