--- nanome/_internal/integration/serialization/export_locations.py
@@ -1,19 +1,0 @@
-from nanome._internal.util.type_serializers import TypeSerializer, StringSerializer, ArraySerializer
-
-
-class _ExportLocations(TypeSerializer):
-    def __init__(self):
-        self.array = ArraySerializer()
-        self.array.set_type(StringSerializer())
-
-    def version(self):
-        return 0
-
-    def name(self):
-        return "ExportLocations"
-
-    def serialize(self, version, value, context):
-        context.write_using_serializer(self.array, value)
-
-    def deserialize(self, version, context):
-        return None
