--- nanome/_internal/integration/serialization/import_smiles.py
@@ -1,30 +1,0 @@
-from nanome._internal.util.type_serializers import TypeSerializer, ArraySerializer, DictionarySerializer, LongSerializer, StringSerializer
-from nanome._internal.structure.serialization import _ComplexSerializer, _AtomSerializer
-
-
-class _ImportSmiles(TypeSerializer):
-    def __init__(self):
-        self.complex_array = ArraySerializer()
-        self.complex_array.set_type(_ComplexSerializer())
-        self.string_array = ArraySerializer()
-        self.string_array.set_type(StringSerializer())
-
-        self.dict = DictionarySerializer()
-        self.dict.set_types(LongSerializer(), _AtomSerializer())
-
-    def version(self):
-        return 0
-
-    def name(self):
-        return "ImportSmiles"
-
-    def serialize(self, version, value, context):
-        subcontext = context.create_sub_context()
-        subcontext.payload["Atom"] = {}
-        subcontext.write_using_serializer(self.complex_array, value)
-        context.write_using_serializer(self.dict, subcontext.payload["Atom"])
-        context.write_bytes(subcontext.to_array())
-
-    def deserialize(self, version, context):
-        strings = context.read_using_serializer(self.string_array)
-        return strings
